recovers deleted JPGs from a file
Since when you "delete" something it's not actually deleted, it's still somewhere in the memory; you just kinda lost its address, one of this week's problems was to write a program that recovers deleted images from a file.

This works by going through every 512 bytes of memory and recognising if it's the start of a new image or a continuation of one. If it's a new image, create a new file to store its data in, else write the data on the currently opened image file.