Identify a person by their DNA sequence
Week 6
Problem set 6
The program takes two command-line arguments, one works as our DNA dataBase with every person with each number of STR chains for each STR.
It then compares the str chains in the DNA we want to identify with each person in our database. If it found a match, it prints its name. Else, it prints "no match".