Solve a mystery using a series of SQL queries
Week 7
Problem set 7
write a series of queries to solve a mystery: sometime at July 28, 2020 on Chamberlin Street a crime took place, a duck was stolen! The goal is to find out who the thief is, where he travelled after committing the crime, and who helped him. 
With nothing to go on but the time and place of the crime, and a huge amount of data like interviews with the witnesses, security camera logs, flights records. I put your detective hat on and took a dive through all that data, examining, and cross-referencing each piece with the other, which, despite the silliness of this assignment, is actually how most detective work is